Added commands to the code and fixed some bugs
Fixed the bug that maybe occured, when you have statistics of the map Icebox and Breeze. There was some wrong parsing coordinates from Google Sheets. Also updated the location of the map breeze on information of this article https://www.washingtonpost.com/video-games/2021/04/26/heres-how-breeze-valorants-newest-map-was-built/ to the country Trinidad and Tobago.